atualizando projeto

+++ projeto_organizador_arquivos_em_andamento/index.js
@@ -2,14 +2,8 @@ const fs = require('node:fs')
 const os = require('node:os')
 const path = require('node:path')
 
-const dirDocuments = 'Documents'
-const dirPictures = 'Pictures'
-const dirDownloads = 'Downloads'
-const dirMusic = 'Music'
-const dirVideos = 'Videos'
-
 const caminhoPastaOrganizadora = path.join(os.homedir(), 'desktop', 'Say_Watcher')
-const caminhoPastaFinal = path.join(os.homedir(), dirPictures, 'Meus_arquivos')
+const caminhoPastaFinal = path.join(os.homedir(), config()[1], 'Meus_arquivos')
 
 const nome = path.basename
 const extensao = path.extname
@@ -29,34 +23,38 @@ function criarPastas() { //verificará a existência das pastas
 }
 criarPastas()
 
-function criarArquivos() { //verificará a existência e criará os arquivos de configuração
+function criarArquivos() { //verificará a existência e criará os arquivos TXT
 
     function leiameTXT() {
     fs.writeFileSync(
             path.join(caminhoPastaOrganizadora, 'LEIAME.config', 'LEIAME.txt'),
             `========================================
        📂 Aplicação de Organização para Arquivos
-    ========================================
+========================================
     
      Bem-vindo(a)!  
-     Você acaba de adquirir, uma ferramenta desenvolvida para facilitar 
-     a organização automática de seus arquivos em seu computador.
+     Você acaba de adquirir, uma ferramenta desenvolvida para facilitar a organização automática de seus arquivos em seu computador.
     
-     ----------------------------------------
-     🔧 Manual de procedimento:
-    
+     - Essa pasta chamada "Say_Watcher" será onde você colocará todos os seus arquivos que serão organizados
+     
+     - Dentro dela, você encontrou essa pasta "LEIAME.config" com dois arquivos, esse que está lendo agora e outro com algumas configurações da sua aplicação
+     
+     - A pasta onde os arquivos ficarão organizados se chama "Meus_arquivos" e, por padrão, será criada no endereço "Documentos" 
+     
+     - Não apague nem renomeie as pastas "Say_Watcher", "LEIAME.config" e "Meus_arquivos" nem apague ou renomeie os arquivos "LEIAME.txt" e ".config.txt"
+     
+     ---------------------------------------------
+        🔧 Manual de procedimento:
+
      ⭐ Atalho??
      ⭐ .config
     
-     - Sua pasta, onde os arquivos serão enviados para organização, se chama "Say_Watcher" e se encontra em sua área de trabalho.
-    
-     - Sua pasta, com os arquivos já organizados fica, por padrão, em seus documentos
-    
      - Mova seus arquivos para sua pasta Say_Watcher e aperte o atalho que você configurou
      
-     - Dentro da pasta Say_Watcher, você encontrará um arquivo oculto chamado .config. Esse arquivo permite algumas mudanças no comportamento dessa aplicação. 
-    
-     ----------------------------------------
+     - Dentro da pasta Say_Watcher, você encontrará essa pasta chamada LEIAME.config. Dentro dela, temos esse arquivo que está lendo agora e o arquivo .config, responsável por alterar alguns processos na sua aplicação.  
+     ❗ - Não deixe de ler o seu arquivo .config - ❗
+     ---------------------------------------------
+
      Obrigado por confiar em nossa aplicação! 🚀`,
             'utf-8'
         )
@@ -65,7 +63,11 @@ function criarArquivos() { //verificará a existência e criará os arquivos de 
     function configTXT() {
     fs.writeFileSync(
             path.join(caminhoPastaOrganizadora, 'LEIAME.config', '.config.txt'),
-            `========================================`,
+            `========================================
+    const dirDocuments = 'Documents'
+    const dirPictures = 'Pictures'
+    const dirMusic = 'Music'
+    const dirVideos = 'Videos'`,
             'utf-8'
         )
     }
@@ -85,19 +87,17 @@ function config() {
     const arquivoDeConfiguracao = fs.readFileSync(path.join(caminhoPastaOrganizadora, 'LEIAME.config', '.config.txt'), 'utf-8')
     const regex = /\{([^\}]+)\}/g
     const opcoesEmString = arquivoDeConfiguracao.matchAll(regex)
-    const opcoesEmArray = [...opcoesEmString].map(valor => valor[1].toLocaleLowerCase().replaceAll(/n/g, ' '))//.replaceAll('ã', 'a'))
+    const opcoesEmArray = [...opcoesEmString].map(valor => valor[1].toLocaleLowerCase().replaceAll(' ', '').replaceAll('ã', 'a'))
 
     return opcoesEmArray
 }
-
-console.log(config())
 
 function mover_e_ValidarArquivos(arquivo) {
 
     if(fs.existsSync(path.join(caminhoPastaFinal, extensao(arquivo), nome(arquivo)))) {
        return fs.copyFileSync(
             path.join(caminhoPastaOrganizadora, nome(arquivo)),
-            path.join(caminhoPastaFinal, extensao(arquivo), nome(arquivo).replace(extensao(arquivo),  `- copia${extensao(arquivo)}`))) 
+            path.join(caminhoPastaFinal, extensao(arquivo), nome(arquivo))) 
     } else {
         return fs.copyFileSync(
             path.join(caminhoPastaOrganizadora, nome(arquivo)),
@@ -116,13 +116,16 @@ arqPastaOrganizadora.forEach(arquivo => {
 
 arrayArquivos.forEach(arquivo => {
 
-        if (!fs.existsSync(path.join(caminhoPastaFinal, extensao(arquivo)))) {
-            fs.mkdirSync(path.join(caminhoPastaFinal, extensao(arquivo)))
+    if(!fs.existsSync(path.join(caminhoPastaFinal, extensao(arquivo)))) {
+        fs.mkdirSync(path.join(caminhoPastaFinal, extensao(arquivo)))
 
-            mover_e_ValidarArquivos(arquivo)
+        mover_e_ValidarArquivos(arquivo)
 
-        } else {
-            mover_e_ValidarArquivos(arquivo)
-        }
+    } else {
+        mover_e_ValidarArquivos(arquivo)
+    }
+
+    if(config()[0] == 'nao' || config()[0] == '') { //opção (copiar ou mover)
         fs.unlinkSync(path.join(caminhoPastaOrganizadora, nome(arquivo)))
-    })
+    }
+})
